ENG-13403: Renamed NIOWriteStream to VoltNIOWriteStream. Separated out TLS encryption related logic into TLSEncryptionAdapter class.
