+++ src/frontend/org/voltcore/network/NIOWriteStreamBase.java
@@ -169,7 +169,7 @@ public abstract class NIOWriteStreamBase {
      * Validate that serialization is accurately reporting the amount of data necessary
      * to serialize the message
      */
-    protected void checkSloppySerialization(ByteBuffer buf, DeferredSerialization ds) {
+    protected static void checkSloppySerialization(ByteBuffer buf, DeferredSerialization ds) {
         if (buf.limit() != buf.capacity()) {
             if (ASSERT_ON) {
                 networkLog.fatal("Sloppy serialization size for message class " + ds);
