+++ src/frontend/org/voltcore/network/TLSNIOWriteStream.java
@@ -18,11 +18,9 @@
 package org.voltcore.network;
 
 import java.io.IOException;
-import java.nio.ByteBuffer;
 import java.nio.channels.GatheringByteChannel;
 import java.util.ArrayList;
 import java.util.Deque;
-import java.util.Iterator;
 import java.util.List;
 import java.util.concurrent.ConcurrentLinkedDeque;
 import java.util.concurrent.ExecutionException;
@@ -30,141 +28,53 @@ import java.util.concurrent.TimeUnit;
 
 import javax.net.ssl.SSLEngine;
 
-import org.voltcore.utils.CoreUtils;
 import org.voltcore.utils.DeferredSerialization;
 import org.voltcore.utils.EstTime;
 import org.voltcore.utils.FlexibleSemaphore;
-import org.voltcore.utils.ssl.SSLBufferEncrypter;
+import org.voltcore.utils.Pair;
 
-import com.google_voltpatches.common.collect.ImmutableList;
-import com.google_voltpatches.common.util.concurrent.ListenableFuture;
-
-import io.netty_voltpatches.buffer.ByteBuf;
 import io.netty_voltpatches.buffer.CompositeByteBuf;
 import io.netty_voltpatches.buffer.Unpooled;
 
-public class TLSNIOWriteStream extends NIOWriteStream {
+public class TLSNIOWriteStream extends VoltNIOWriteStream {
 
     private final ConcurrentLinkedDeque<ExecutionException> m_exceptions = new ConcurrentLinkedDeque<>();
     private final ConcurrentLinkedDeque<EncryptFrame> m_encrypted = new ConcurrentLinkedDeque<>();
     private final FlexibleSemaphore m_inFlight = new FlexibleSemaphore(1);
 
     private final CompositeByteBuf m_outbuf;
-    private final CipherExecutor m_ce;
-    private final SSLEngine m_sslEngine;
-    private final SSLBufferEncrypter m_encrypter;
-    private final EncryptionGateway m_ecryptgw = new EncryptionGateway();
     private int m_queuedBytes = 0;
+    private final TLSEncryptionAdapter m_encryptAdapter;
 
     public TLSNIOWriteStream(Connection port, Runnable offBackPressureCallback,
             Runnable onBackPressureCallback, QueueMonitor monitor,
             SSLEngine engine, CipherExecutor cipherExecutor) {
         super(port, offBackPressureCallback, onBackPressureCallback, monitor);
-        m_sslEngine = engine;
-        m_ce = cipherExecutor;
         m_outbuf = Unpooled.compositeBuffer();
-        m_encrypter = new SSLBufferEncrypter(engine);
-    }
-
-    /**
-     * this values may change if a TLS session renegotiates its cipher suite
-     */
-    private int applicationBufferSize() {
-        return m_sslEngine.getSession().getApplicationBufferSize();
-    }
-
-    /**
-     * this values may change if a TLS session renegotiates its cipher suite
-     */
-    private int packetBufferSize() {
-        return m_sslEngine.getSession().getPacketBufferSize();
+        m_encryptAdapter = new TLSEncryptionAdapter(port, engine, cipherExecutor, m_encrypted);
     }
 
     @Override
     int serializeQueuedWrites(NetworkDBBPool pool) throws IOException {
         checkForGatewayExceptions();
 
-        final int frameMax = Math.min(CipherExecutor.FRAME_SIZE, applicationBufferSize());
-        int processedWrites = 0;
         final Deque<DeferredSerialization> oldlist = getQueuedWrites();
         if (oldlist.isEmpty()) return 0;
 
-        ByteBuf accum = m_ce.allocator().buffer(frameMax).clear();
+        Pair<Integer, Integer> processedWrites = m_encryptAdapter.encryptBuffers(oldlist);
 
-        DeferredSerialization ds = null;
-        int bytesQueued = 0;
-        int frameMsgs = 0;
-        while ((ds = oldlist.poll()) != null) {
-            ++processedWrites;
-            final int serializedSize = ds.getSerializedSize();
-            if (serializedSize == DeferredSerialization.EMPTY_MESSAGE_LENGTH) continue;
-            // pack as messages you can inside a TLS frame before you send it to
-            // the encryption gateway
-            if (serializedSize > frameMax) {
-                // frames may contain only one or more whole messages, or only
-                // partial parts of one message. a message may not contain whole
-                // messages and an incomplete partial fragment of one
-                if (accum.writerIndex() > 0) {
-                    m_ecryptgw.offer(new EncryptFrame(accum, frameMsgs));
-                    frameMsgs = 0;
-                    bytesQueued += accum.writerIndex();
-                    accum = m_ce.allocator().buffer(frameMax).clear();
-                }
-                ByteBuf big = m_ce.allocator().buffer(serializedSize).writerIndex(serializedSize);
-                ByteBuffer jbb = big.nioBuffer();
-                ds.serialize(jbb);
-                checkSloppySerialization(jbb, ds);
-                bytesQueued += big.writerIndex();
-                m_ecryptgw.offer(new EncryptFrame(big, 1));
-                frameMsgs = 0;
-                continue;
-            } else if (accum.writerIndex() + serializedSize > frameMax) {
-                m_ecryptgw.offer(new EncryptFrame(accum, frameMsgs));
-                frameMsgs = 0;
-                bytesQueued += accum.writerIndex();
-                accum = m_ce.allocator().buffer(frameMax).clear();
-            }
-            ByteBuf packet = accum.slice(accum.writerIndex(), serializedSize);
-            ByteBuffer jbb = packet.nioBuffer();
-            ds.serialize(jbb);
-            checkSloppySerialization(jbb, ds);
-            accum.writerIndex(accum.writerIndex()+serializedSize);
-            ++frameMsgs;
-        }
-        if (accum.writerIndex() > 0) {
-            m_ecryptgw.offer(new EncryptFrame(accum, frameMsgs));
-            bytesQueued += accum.writerIndex();
-        } else {
-            accum.release();
-        }
-        updateQueued(bytesQueued, true);
-        return processedWrites;
+        updateQueued(processedWrites.getSecond(), true);
+        return processedWrites.getFirst();
+    }
+
+    public void waitForPendingEncrypts() throws IOException {
+        m_encryptAdapter.waitForPendingEncrypts();
     }
 
     @Override
     public void updateQueued(int queued, boolean noBackpressureSignal) {
         super.updateQueued(queued, noBackpressureSignal);
         m_queuedBytes += queued;
-    }
-
-    void waitForPendingEncrypts() throws IOException {
-        boolean acquired;
-
-        do {
-            int waitFor = 1 - m_inFlight.availablePermits();
-            acquired = waitFor == 0;
-            for (int i = 0; i < waitFor && !acquired; ++i) {
-                checkForGatewayExceptions();
-                try {
-                    acquired = m_inFlight.tryAcquire(1, TimeUnit.SECONDS);
-                    if (acquired) {
-                        m_inFlight.release();
-                    }
-                } catch (InterruptedException e) {
-                    throw new IOException("interrupted while waiting for pending encrypts", e);
-                }
-            }
-        } while (!acquired);
     }
 
     private final List<EncryptFrame> m_partial = new ArrayList<>();
@@ -225,7 +135,7 @@ public class TLSNIOWriteStream extends NIOWriteStream {
     @Override
     synchronized public boolean isEmpty() {
         return m_queuedWrites.isEmpty()
-            && m_ecryptgw.isEmpty()
+            && m_encryptAdapter.isEmpty()
             && m_encrypted.isEmpty()
             && m_partialSize == 0
             && !m_outbuf.isReadable();
@@ -299,7 +209,7 @@ public class TLSNIOWriteStream extends NIOWriteStream {
                 .append("isEmpty()=").append(isEmpty())
                 .append(", encrypted.isEmpty()=").append(m_encrypted.isEmpty())
                 .append(", exceptions.isEmpty()=").append(m_exceptions.isEmpty())
-                .append(", gateway=").append(m_ecryptgw.dumpState())
+                .append(", gateway=").append(m_encryptAdapter.dumpState())
                 .append(", inFligth=").append(m_inFlight.availablePermits())
                 .append(", outbuf.readableBytes()=").append(m_outbuf.readableBytes())
                 .append("]").toString();
@@ -334,7 +244,7 @@ public class TLSNIOWriteStream extends NIOWriteStream {
                 }
             }
 
-            m_ecryptgw.die();
+            m_encryptAdapter.shutdown();
 
             EncryptFrame frame = null;
             while ((frame = m_encrypted.poll()) != null) {
@@ -357,165 +267,4 @@ public class TLSNIOWriteStream extends NIOWriteStream {
             m_inFlight.release();
         }
     }
-
-    /**
-     * Construct used to serialize all the encryption tasks for this stream.
-     * it takes an encryption request offer, divides it into chunks that
-     * can be handled wholly by SSLEngine wrap, and queues all the
-     * encrypted frames to the m_encrypted queue. All faults are queued
-     * to the m_exceptions queue
-     */
-    class EncryptionGateway implements Runnable {
-        private final ConcurrentLinkedDeque<EncryptFrame> m_q = new ConcurrentLinkedDeque<>();
-        private final int COALESCE_THRESHOLD = CipherExecutor.FRAME_SIZE - 4096;
-
-        synchronized void offer(EncryptFrame frame) throws IOException {
-            final boolean wasEmpty = m_q.isEmpty();
-
-            List<EncryptFrame> chunks = frame.chunked(
-                    Math.min(CipherExecutor.FRAME_SIZE, applicationBufferSize()));
-
-            m_q.addAll(chunks);
-            m_inFlight.reducePermits(chunks.size());
-
-            if (wasEmpty) {
-                submitSelf();
-            }
-        }
-
-        /**
-         * Encryption takes time and the likelihood that some or more encrypt frames
-         * are queued up behind the recently completed frame encryption is high. This
-         * takes queued up small frames and coalesces them into a bigger frame
-         */
-        private void coalesceEncryptFrames() {
-            EncryptFrame head = m_q.peek();
-            if (head == null || head.chunks > 1 || head.bb.readableBytes() > COALESCE_THRESHOLD) {
-                return;
-            }
-
-            m_q.poll();
-            ByteBuf bb = head.bb;
-            int msgs = head.msgs;
-            int released = 0;
-
-            head = m_q.peek();
-            while (head != null && head.chunks == 1 && head.bb.readableBytes() <= bb.writableBytes()) {
-                m_q.poll();
-                bb.writeBytes(head.bb, head.bb.readableBytes());
-                head.bb.release();
-                ++released;
-                msgs += head.msgs;
-                head = m_q.peek();
-            }
-            m_q.push(new EncryptFrame(bb, 0, msgs));
-            if (released > 0) {
-                m_inFlight.release(released);
-            }
-        }
-
-        synchronized int die() {
-            int toUnqueue = 0;
-            EncryptFrame ef = null;
-            while ((ef = m_q.poll()) != null) {
-                toUnqueue += ef.frame.readableBytes();
-                if (ef.isLast()) {
-                    ef.bb.release();
-                }
-            }
-            return toUnqueue;
-        }
-
-        String dumpState() {
-            return new StringBuilder(256).append("EncryptionGateway[")
-                    .append("q.isEmpty()=").append(m_q.isEmpty())
-                    .append(", partialSize=").append(m_partialSize)
-                    .append("]").toString();
-        }
-
-        public Iterator<EncryptFrame> iterator() {
-            return ImmutableList.copyOf(m_q).iterator();
-        }
-
-        @Override
-        public void run() {
-            EncryptFrame frame = m_q.peek();
-            if (frame == null) return;
-
-            ByteBuffer src = frame.frame.nioBuffer();
-            ByteBuf encr = m_ce.allocator().ioBuffer(packetBufferSize()).writerIndex(packetBufferSize());
-            ByteBuffer dest = encr.nioBuffer();
-
-            try {
-                m_encrypter.tlswrap(src, dest);
-            } catch (TLSException e) {
-                m_inFlight.release();
-                encr.release();
-                m_exceptions.offer(new ExecutionException("failed to encrypt frame", e));
-                networkLog.error("failed to encrypt frame", e);
-                m_port.enableWriteSelection();
-                return;
-            }
-            assert !src.hasRemaining() : "encryption wrap did not consume the whole source buffer";
-            int delta = dest.limit() - frame.frame.readableBytes();
-            encr.writerIndex(dest.limit());
-
-            if (!m_isShutdown) {
-                m_encrypted.offer(frame.encrypted(delta, encr));
-                /*
-                 * All interactions with write stream must be protected
-                 * with a lock to ensure that interests ops are consistent with
-                 * the state of writes queued to the stream. This prevent
-                 * lost queued writes where the write is queued
-                 * but the write interest op is not set.
-                 */
-                if (frame.isLast()) {
-                    m_port.enableWriteSelection();
-                }
-            } else {
-                encr.release();
-                return;
-            }
-            synchronized(this) {
-                m_q.poll();
-                if (frame.isLast()) {
-                    frame.bb.release();
-                }
-                m_inFlight.release();
-                coalesceEncryptFrames();
-                if (m_q.peek() != null && !m_isShutdown) {
-                    submitSelf();
-                }
-            }
-        }
-
-        boolean isEmpty() {
-            return m_q.isEmpty();
-        }
-
-        void submitSelf() {
-            ListenableFuture<?> fut = m_ce.submit(this);
-            fut.addListener(new ExceptionListener(fut), CoreUtils.LISTENINGSAMETHREADEXECUTOR);
-        }
-    }
-
-    class ExceptionListener implements Runnable {
-        private final ListenableFuture<?> m_fut;
-        private ExceptionListener(ListenableFuture<?> fut) {
-            m_fut = fut;
-        }
-        @Override
-        public void run() {
-            if (!m_isShutdown) return;
-            try {
-                m_fut.get();
-            } catch (InterruptedException notPossible) {
-            } catch (ExecutionException e) {
-                m_inFlight.release();
-                networkLog.error("unexpect fault occurred in encrypt task", e.getCause());
-                m_exceptions.offer(e);
-            }
-        }
-    }
-
 }
